backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Validation and serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# File handling
aiofiles>=23.2.1
Pillow>=10.0.0

# Email
aiosmtplib>=3.0.1
jinja2>=3.1.3

# Excel export
openpyxl>=3.1.2
pandas>=2.2.0

# Security
python-dotenv>=1.0.0

# CORS
starlette>=0.35.1

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
